Not saving session information after registration so that user will have to login after registration. this is safeguard for otp validation. removed bug from attachprescription.java destroy() method

+++ Patient/src/com/medikeen/patient/AttachPrescription.java
@@ -164,8 +164,10 @@ public class AttachPrescription extends Activity implements OnClickListener {
 		File dir = new File(GridViewDemo_ImagePath);
 		if (dir.isDirectory()) {
 			String[] children = dir.list();
-			for (int i = 0; i < children.length; i++) {
-				new File(dir, children[i]).delete();
+			if(children!= null) {
+				for (int i = 0; i < children.length; i++) {
+					new File(dir, children[i]).delete();
+				}
 			}
 		}
 	}
